Move some case classes to a different package
