+++ test/mocks/application/DocumentRootInMemoryRepository.scala
@@ -22,10 +22,9 @@ import model.Commands._
 import model.EvaluationResults.AssessmentRuleCategoryResult
 import model.Exceptions.ApplicationNotFound
 import model.OnlineTestCommands.OnlineTestApplication
-import model.persisted.Phase1TestProfile
+import model.persisted.{ApplicationForDiversityReport, ApplicationForNotification, Phase1TestProfile}
 import model._
 import model.command._
-import model.persisted.ApplicationForNotification
 import model.report._
 import org.joda.time.{DateTime, LocalDate}
 import repositories.application.GeneralApplicationRepository
